LayoutTests/editing/selection/ios/select-all-in-readonly-input-does-not-overflow-expected.txt

This test verifies that selection views are clipped to the bounds of a readonly input. To manually run the test, tap the 'Select' button, and then tap the input.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS selectionRects.length is 1
PASS input.getBoundingClientRect().width is >= selectionRects[0].width
PASS successfullyParsed is true

TEST COMPLETE
